requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pydobot==1.3.2
pyserial==3.4
tinydb==4.8.0
Werkzeug==3.0.2